+++ src/gamspy/_model.py
@@ -20,6 +20,7 @@ import gamspy._validation as validation
 import gamspy.utils as utils
 from gamspy._backend.backend import backend_factory
 from gamspy._model_instance import ModelInstance
+from gamspy._symbols.symbol import Symbol
 from gamspy.exceptions import GamspyException, ValidationError
 
 if TYPE_CHECKING:
@@ -766,7 +767,7 @@ class Model:
 
         return model_str
 
-    def toGams(self, path: str) -> None:
+    def toGams(self, path: str, skip_endogenous_records: bool = False) -> None:
         """
         Generates GAMS model under path/<model_name>.gms
 
@@ -780,13 +781,21 @@ class Model:
         def sort_names(name):
             PRECEDENCE = {
                 gp.Set: 1,
-                gp.Alias: 2,
+                gp.Alias: 1,
                 gp.Parameter: 3,
                 gp.Variable: 4,
                 gp.Equation: 5,
             }
 
-            return PRECEDENCE[type(self.container[name])]
+            symbol = self.container[name]
+            precedence = PRECEDENCE[type(symbol)]
+
+            if isinstance(symbol, gp.Set) and any(
+                not isinstance(elem, str) for elem in symbol.domain
+            ):
+                precedence = 2
+
+            return precedence
 
         all_symbols = []
         definitions = []
@@ -798,22 +807,47 @@ class Model:
                 if symbol not in all_symbols:
                     all_symbols.append(symbol)
 
+        if self._matches:
+            for equation, variable in self._matches.items():
+                if (
+                    equation.name not in all_symbols
+                    and not skip_endogenous_records
+                ):
+                    symbols = equation._definition._find_all_symbols()
+                    for symbol in symbols:
+                        if symbol not in all_symbols:
+                            all_symbols.append(symbol)
+
+                    if equation.name not in all_symbols:
+                        all_symbols.append(equation.name)
+
+                    definitions.append(equation._definition.getDeclaration())
+
+                if (
+                    variable.name not in all_symbols
+                    and not skip_endogenous_records
+                ):
+                    for elem in variable.domain:
+                        if not isinstance(elem, Symbol):
+                            continue
+
+                        elem_path = validation.get_domain_path(elem)
+                        for name in elem_path:
+                            if name not in all_symbols:
+                                all_symbols.append(name)
+
+                    if variable.name not in all_symbols:
+                        all_symbols.append(variable.name)
+
         all_needed_symbols = sorted(all_symbols, key=sort_names)
-        loadable_symbols = [
-            name
-            for name in all_needed_symbols
-            if not isinstance(self.container[name], gp.Alias)
-        ]
         gdx_path = os.path.join(path, self.name + "_data.gdx")
-        self.container.write(gdx_path)
+        self.container.write(gdx_path, all_needed_symbols)
 
         strings = [
             self.container[name].getDeclaration()
             for name in all_needed_symbols
         ]
-        strings.append(f"$gdxIn {os.path.abspath(gdx_path)}")
-        strings.append(f'$loadDC {",".join(loadable_symbols)}')
-        strings.append("$gdxIn")
+        strings.append(f"$gdxLoadAll {os.path.abspath(gdx_path)}")
         strings += definitions
         strings.append(self.getDeclaration())
         solve_string = self._generate_solve_string()